+++ viewer/lil-gui-master/scripts/homepage.js
@@ -4,11 +4,11 @@ import hljs from 'highlight.js';
 import hbs from 'handlebars';
 import { execSync } from 'child_process';
 
-import pkg from '../package.json';
-import jsdocData from './api';
+import pkg from './package.js';
+import jsdocData from './api.js';
 
 const OUTPUT = 'index.html';
-const TEMPLATE = 'scripts/homepage.hbs.html';
+const TEMPLATE = 'homepage/homepage.hbs.html';
 const README = 'README.md';
 const GUIDE = 'Guide.md';
 const MIGRATING = 'Migrating.md';
@@ -22,7 +22,7 @@ console.time( 'homepage' );
 let readme = read( README );
 
 // remove homepage link
-readme = readme.replace( '[**Homepage**](https://lil-gui.georgealways.com/) • ', '' );
+readme = readme.replace( '[**Homepage**](https://lil-gui.georgealways.com/) •', '' );
 
 // Guide.md
 // -----------------------------------------------------------------------------