+++ viewer/lil-gui-master/src/FunctionController.js
@@ -14,10 +14,11 @@ export default class FunctionController extends Controller {
 		this.$button.addEventListener( 'click', e => {
 			e.preventDefault();
 			this.getValue().call( this.object );
+			this._callOnChange();
 		} );
 
 		// enables :active pseudo class on mobile
-		this.$button.addEventListener( 'touchstart', () => { } );
+		this.$button.addEventListener( 'touchstart', () => {}, { passive: true } );
 
 		this.$disable = this.$button;