+++ viewer/lil-gui-master/tests/on-change-controller.js
@@ -1,5 +1,7 @@
 import assert from 'assert';
-import GUI from '..';
+import GUI from '../dist/lil-gui.esm.min.js';
+
+import CallTracker from './utils/CallTracker.js';
 
 export default () => {
 
@@ -7,24 +9,31 @@ export default () => {
 
 	const controller = gui.add( { x: 0 }, 'x' );
 
-	let tracker = new assert.CallTracker();
+	let tracker = new CallTracker();
 
-	let handler, _args, _this;
-
-	handler = tracker.calls( function( ...args ) {
-		_this = this;
-		_args = args;
-	}, 1 );
-
-	controller.onChange( handler );
+	controller.onChange( tracker.handler );
 
 	const value = 9;
 	controller.setValue( value );
 
-	assert.strictEqual( controller._onChange, handler, 'onChange: sets _onChange' );
-	assert.strictEqual( _this, controller, 'onChange: this is bound to controller in handler' );
-	assert.deepEqual( _args, [ value ], 'onChange: new value is the first and only argument' );
+	assert.strictEqual( controller._onChange, tracker.handler, 'onChange: sets _onChange' );
+	assert.strictEqual( tracker.lastThis, controller, 'onChange: this is bound to controller in handler' );
+	assert.deepEqual( tracker.lastArgs, [ value ], 'onChange: new value is the first and only argument' );
 
-	tracker.verify();
+	assert.strictEqual( tracker.numCalls, 1 );
+
+	controller.setValue( value );
+
+	assert.strictEqual( tracker.numCalls, 1, 'redundant calls to setValue do not trigger onChange' );
+
+	// function controller should call onChange after click
+
+	tracker = new CallTracker();
+
+	const functionController = gui.add( { fnc() {} }, 'fnc' );
+	functionController.onChange( tracker.handler );
+	functionController.$button.$callEventListener( 'click' );
+
+	assert.strictEqual( tracker.numCalls, 1 );
 
 };