+++ viewer/lil-gui-master/tests/utils/shim.js
@@ -41,6 +41,7 @@ class Element extends EventTarget {
 	}
 	appendChild() {}
 	removeChild() {}
+	replaceChildren() {}
 	insertBefore() {}
 	setAttribute() {}
 	removeAttribute() {}
@@ -79,6 +80,7 @@ function createElement( tag ) {
 }
 
 global.window = new EventTarget();
+global.window.matchMedia = () => { return { matches: true }; };
 global.requestAnimationFrame = fnc => setTimeout( fnc, 100 / 6 );
 global.cancelAnimationFrame = id => clearTimeout( id );